requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
pyotp==2.8.0
pytz==2024.1
sqlparse==0.4.4
